chore: refactor useSendChatMessage
