+++ web-client/app/_hooks/useChatMessages.ts
@@ -27,7 +27,7 @@ const useChatMessages = (offset = 0) => {
   const updateConvoHasMore = useSetAtom(updateConversationHasMoreAtom);
 
   useEffect(() => {
-    const hasMore = convoStates[currentConvo.id]?.hasMore ?? true;
+    const hasMore = convoStates[currentConvo.id ?? ""]?.hasMore ?? true;
     if (!hasMore) return;
 
     const getMessages = async () => {
@@ -40,7 +40,7 @@ const useChatMessages = (offset = 0) => {
         }
         parseMessages(data ?? []).then((newMessages) => {
           addOldChatMessages(newMessages);
-          updateConvoHasMore(currentConvo.id, false);
+          updateConvoHasMore(currentConvo.id ?? "", false);
           setLoading(false);
         });
       });
@@ -57,7 +57,7 @@ const useChatMessages = (offset = 0) => {
   return {
     loading: loading,
     error: undefined,
-    hasMore: convoStates[currentConvo.id]?.hasMore ?? true,
+    hasMore: convoStates[currentConvo.id ?? ""]?.hasMore ?? true,
   };
 };
